requirements.txt
flask>=2.3.0
requests>=2.31.0
beautifulsoup4>=4.12.0
schedule>=1.2.0
aiohttp>=3.8.0
pytz>=2024.1
discord.py>=2.3.2
playwright>=1.40.0
